+++ index.html
@@ -226,7 +226,6 @@
       </div>
     </section>
   </template>
-  <script defer src="js/main.js"></script>
-<script defer src="js/functions.js"></script>
+  <script defer src="js/main.js" type="module"></script>
 </body>
 </html>
